+++ CSharp/SchedulingClients.UserControls/JobBuilderClientControl.xaml.cs
@@ -9,14 +9,14 @@ namespace SchedulingClients.UserControls
 {
     public partial class JobBuilderClientControl : UserControl
     {
-        private RoadmapClient roadmapClient = null;
+        private MapClient roadmapClient = null;
 
         public JobBuilderClientControl()
         {
             InitializeComponent();
         }
 
-        public void Configure(RoadmapClient client)
+        public void Configure(MapClient client)
         {
             roadmapClient = client;
         }
