--- CSharp/SchedulingClients.UserControls/MappingKeyDialog.xaml
@@ -1,35 +1,0 @@
-﻿<Window x:Class="SchedulingClients.UserControls.MappingKeyDialog"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:SchedulingClients.UserControls"
-        mc:Ignorable="d"
-        xmlns:system="clr-namespace:System;assembly=mscorlib"
-        Title="Mapping Key" Height="300" Width="300">
-    <Window.Resources>
-        <system:Double x:Key="Text">14</system:Double>
-        <Style TargetType="{x:Type Control}" x:Key="BaseControlStyle">
-            <Setter Property="FontFamily" Value="Calibri" />
-            <Setter Property="FontSize" Value="{StaticResource Text}" />
-            <Setter Property="Margin" Value="6" />
-        </Style>
-        <Style TargetType="{x:Type TextBlock}" x:Key="BaseTextBlockStyle">
-            <Setter Property="FontFamily" Value="Calibri" />
-            <Setter Property="FontSize" Value="{StaticResource Text}" />
-            <Setter Property="Margin" Value="6" />
-        </Style>
-        <Style TargetType="{x:Type Button}" BasedOn="{StaticResource BaseControlStyle}">
-            <Setter Property="Height" Value="23" />
-            <Setter Property="Width" Value="75" />
-        </Style>
-    </Window.Resources>
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="*" />
-            <RowDefinition Height="Auto" />
-        </Grid.RowDefinitions>
-        <TextBlock Text="{Binding}" TextWrapping="Wrap" />
-        <Button Grid.Row="1" HorizontalAlignment="Right" x:Name="okButton" Click="okButton_Click">OK</Button>
-    </Grid>
-</Window>
