--- CSharp/SchedulingClients.UserControls/MappingKeyDialog.xaml.cs
@@ -1,32 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
-
-namespace SchedulingClients.UserControls
-{
-    /// <summary>
-    /// Interaction logic for MappingKeyDialog.xaml
-    /// </summary>
-    public partial class MappingKeyDialog : Window
-    {
-        public MappingKeyDialog()
-        {
-            InitializeComponent();
-        }
-
-        private void okButton_Click(object sender, RoutedEventArgs e)
-        {
-            DialogResult = true;
-        }
-    }
-}
