--- CSharp/SchedulingClients.UserControls/RoadmapClientControl.xaml
@@ -1,56 +1,0 @@
-﻿<UserControl x:Class="SchedulingClients.UserControls.RoadmapClientControl"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:local="clr-namespace:SchedulingClients.UserControls"
-             xmlns:toolkit="http://schemas.xceed.com/wpf/xaml/toolkit"
-             mc:Ignorable="d"
-             d:DesignHeight="300" d:DesignWidth="300">
-    <UserControl.Resources>
-        <local:ServiceTypeArrayToStringConverter x:Key="serviceTypeConverter" />
-    </UserControl.Resources>
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="*" />
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="*" />
-            <RowDefinition Height="Auto" />
-        </Grid.RowDefinitions>
-        <local:LabelContentControl Grid.Row="1" Header="Get All Node Data()">
-            <Button x:Name="getAllNodeDataButton" Click="getAllNodeDataButton_Click">Refresh</Button>
-        </local:LabelContentControl>
-        <DataGrid x:Name="nodeDataDataGrid" Grid.Row="2" AutoGenerateColumns="False" ColumnWidth="*"
-                  Width="{Binding Path=ActualWidth, RelativeSource={RelativeSource Mode=FindAncestor, AncestorType=Control, AncestorLevel=1}}">
-            <DataGrid.Columns>
-                <DataGridTextColumn Header="Alias" Binding="{Binding Alias}" />
-                <DataGridTextColumn Header="MapItemId" Binding="{Binding MapItemId}" />
-                <DataGridTextColumn Header="X (m)" Binding="{Binding X, StringFormat={}{0:N3}}" />
-                <DataGridTextColumn Header="Y (m)" Binding="{Binding Y, StringFormat={}{0:N3}}" />
-                <DataGridTextColumn Header="Heading (rad)" Binding="{Binding HeadingRad, StringFormat={}{0:N2}}" />
-                <DataGridTextColumn Header="Services" Binding="{Binding Services, Converter={StaticResource ResourceKey=serviceTypeConverter}}" />
-            </DataGrid.Columns>
-        </DataGrid>
-        <local:LabelContentControl Grid.Row="3" Header="Get All Move Data()">
-            <Button x:Name="getAllMoveDataButton" Click="getAllMoveDataButton_Click">Refresh</Button>
-        </local:LabelContentControl>
-        <DataGrid x:Name="moveDataDataGrid" Grid.Row="4" AutoGenerateColumns="False" ColumnWidth="*"
-                  Width="{Binding Path=ActualWidth, RelativeSource={RelativeSource Mode=FindAncestor, AncestorType=Control, AncestorLevel=1}}">
-            <DataGrid.Columns>
-                <DataGridTextColumn Header="Alias" Binding="{Binding Alias}" />
-                <DataGridTextColumn Header="Id" Binding="{Binding Id}" />
-                <DataGridTextColumn Header="SourceId" Binding="{Binding SourceId}" />
-                <DataGridTextColumn Header="DestinationId" Binding="{Binding DestinationId}" />
-                <DataGridTextColumn Header="Length (m)" Binding="{Binding Length, StringFormat={}{0:N3}}" />
-            </DataGrid.Columns>
-        </DataGrid>
-        <local:LabelContentControl Grid.Row="5" Header="Get Trajectory">
-            <StackPanel Orientation="Horizontal">
-                <toolkit:IntegerUpDown x:Name="trajectoryIdUpDown" Minimum="0" DefaultValue="0" DisplayDefaultValueOnEmptyText="True" />
-                <Button x:Name="getTrajectoryButton" Click="getTrajectoryButton_Click">Get</Button>
-            </StackPanel>
-        </local:LabelContentControl>
-    </Grid>
-</UserControl>
